Finally added support for viewing webgestalt images. In addition, modified demo case and set it such that it can never ever be delelted by celery

+++ analysis/tasks.py
@@ -126,6 +126,10 @@ def clean_database_and_analysis():
     for analysis in Analysis.objects.all():
         time_threshold = timezone.now() - timedelta(hours=24)
         # no need to delete folders individually as we have a signal that handles that
+        
+        # hash id of demo case, never delete this
+        if analysis.sha_hash == "4684d5ebb2cc6f149da60b9e59055087dbeafe1d":
+            continue
         print(analysis)
         print(analysis.modified)
         Analysis.objects.filter(modified__lt=time_threshold).delete()
